+++ src/mlpack/methods/ann/layer/convolution.hpp
@@ -76,7 +76,7 @@ class Convolution
               const size_t padH = 0,
               const size_t inputWidth = 0,
               const size_t inputHeight = 0,
-              const std::string paddingType = "None");
+              const std::string& paddingType = "None");
 
   /**
    * Create the Convolution object using the specified number of input maps,
@@ -104,11 +104,11 @@ class Convolution
               const size_t kernelHeight,
               const size_t strideWidth,
               const size_t strideHeight,
-              const std::tuple<size_t, size_t> padW,
-              const std::tuple<size_t, size_t> padH,
+              const std::tuple<size_t, size_t>& padW,
+              const std::tuple<size_t, size_t>& padH,
               const size_t inputWidth = 0,
               const size_t inputHeight = 0,
-              const std::string paddingType = "None");
+              const std::string& paddingType = "None");
 
   /*
    * Set the weight and bias term.